refactor: MERGE 승기님 혼자 모드 게임 + 성철님 혼자 모드 UI
