--- Acha/Acha/Presentation/TabBar/Home/SingleGamePlayViewModel.swift
@@ -1,19 +1,0 @@
-//
-//  SingleGamePlayViewModel.swift
-//  Acha
-//
-//  Created by hong on 2022/11/15.
-//
-
-import Foundation
-import RxSwift
-
-final class SingleGamePlayViewModel {
-    
-    private let coordinator: Coordinator
-    
-    init(coordinator: Coordinator) {
-        self.coordinator = coordinator
-    }
-
-}
